simplecommit

+++ src/main/java/com/potitech/potifood/domain/service/CrudCozinhaService.java
@@ -15,7 +15,7 @@ import com.potitech.potifood.domain.repository.CozinhaRepository;
 import com.potitech.potifood.domain.repository.RestauranteRepository;
 
 @Service
-public class CadastroCozinhaService {
+public class CrudCozinhaService {
 
 	@Autowired
 	private CozinhaRepository cozinhaRepository;
@@ -26,9 +26,29 @@ public class CadastroCozinhaService {
 	public Cozinha salvar(Cozinha cozinha) {
 		return cozinhaRepository.createOrUpdate(cozinha);
 	}
+	
+	public Cozinha atualizar(Long id, Cozinha cozinha) {
+		Cozinha cozinhaAtual = cozinhaRepository.findById(id);
+		
+		if(cozinhaAtual == null) {
+			throw new EntidadeNaoEncontradaException("Não foi encontrado nenhum registro de cozinha com código " + id + " para ser atualizado.");
+		}
+		
+		cozinhaAtual.setNome(cozinha.getNome());
+		
+		cozinha = cozinhaRepository.createOrUpdate(cozinhaAtual);
+		
+		return cozinha;
+	}
 
 	public Cozinha buscar(Long id) {
-		return cozinhaRepository.findById(id);
+		Cozinha cozinha = cozinhaRepository.findById(id);
+		
+		if(cozinha == null) {
+			throw new EntidadeNaoEncontradaException(Cozinha.class, id);	
+		}
+		
+		return cozinha;
 	}
 
 	public List<Cozinha> listar() {
@@ -36,16 +56,22 @@ public class CadastroCozinhaService {
 	}
 
 	public void remover(Long id) {
-		List<Restaurante> restaurantes = restauranteRepository.findByCozinha(id);
+		Cozinha cozinha = cozinhaRepository.findById(id);
+		
+		if(cozinha == null) {
+			throw new EntidadeNaoEncontradaException(Cozinha.class, id);
+		}
+		
+		List<Restaurante> restaurantes = cozinha.getRestaurantes();
 
 		try {
 			for (Restaurante restaurante : restaurantes) {
-				restauranteRepository.delete(restaurante);
+				restauranteRepository.delete(restaurante.getId());
 			}
 			cozinhaRepository.delete(id);
-		} catch (EmptyResultDataAccessException e) {
+		} catch (EmptyResultDataAccessException e) { //caso da cozinha não existir
 			throw new EntidadeNaoEncontradaException("Não existe um cadastro de cozinha com código " + id);
-		} catch (DataIntegrityViolationException e) {
+		} catch (DataIntegrityViolationException e) { //caso de chave estrangeira
 			throw new ErroAoRemoverException("Cozinha de código " + id + " não pode ser removida.");
 		}
 	}
