+++ src/main/java/com/potitech/potifood/infrastructure/repository/EstadoRepositoryImpl.java
@@ -6,21 +6,21 @@ import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.transaction.Transactional;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 
 import com.potitech.potifood.domain.model.entities.Estado;
 import com.potitech.potifood.domain.repository.EstadoRepository;
 
-
 @Component
-public class EstadoRepositoryImpl implements EstadoRepository{
+public class EstadoRepositoryImpl implements EstadoRepository {
 
 	@PersistenceContext
 	private EntityManager entityManager;
-	
+
 	@Override
 	public List<Estado> list() {
-		return entityManager.createQuery("from Estado", Estado.class).getResultList();	
+		return entityManager.createQuery("from Estado", Estado.class).getResultList();
 	}
 
 	@Override
@@ -36,7 +36,13 @@ public class EstadoRepositoryImpl implements EstadoRepository{
 
 	@Override
 	@Transactional
-	public void delete(Estado estado) {
+	public void delete(Long id) {
+		Estado estado = findById(id);
+
+		if (estado == null) {
+			throw new EmptyResultDataAccessException(1);
+		}
+
 		entityManager.remove(estado);
 	}
